move test/testdata => faker/bitfaker/bitcoin.go
