+++ bitstamp/bitstamp_test.go
@@ -5,9 +5,9 @@ import (
 	"time"
 
 	exchange "github.com/openware/irix"
+	"github.com/openware/irix/faker/bitfaker"
 	"github.com/openware/irix/portfolio/banking"
 	"github.com/openware/irix/portfolio/withdraw"
-	testdata "github.com/openware/irix/test"
 	"github.com/openware/pkg/asset"
 	"github.com/openware/pkg/common"
 	"github.com/openware/pkg/currency"
@@ -463,7 +463,7 @@ func TestWithdraw(t *testing.T) {
 		Currency:    currency.BTC,
 		Description: "WITHDRAW IT ALL",
 		Crypto: withdraw.CryptoRequest{
-			Address: testdata.BitcoinTestAddress,
+			Address: bitfaker.BitcoinTestAddress,
 		},
 	}
