+++ btse/btse_test.go
@@ -10,8 +10,8 @@ import (
 
 	exchange "github.com/openware/irix"
 	"github.com/openware/irix/config"
+	"github.com/openware/irix/faker/bitfaker"
 	"github.com/openware/irix/sharedtestvalues"
-	testdata "github.com/openware/irix/test"
 	"github.com/openware/pkg/asset"
 	"github.com/openware/pkg/common"
 	"github.com/openware/pkg/currency"
@@ -581,7 +581,7 @@ func TestCancelExchangeOrder(t *testing.T) {
 		"-")
 	var orderCancellation = &order.Cancel{
 		ID:            "b334ecef-2b42-4998-b8a4-b6b14f6d2671",
-		WalletAddress: testdata.BitcoinTestAddress,
+		WalletAddress: bitfaker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 		AssetType:     asset.Spot,
@@ -613,7 +613,7 @@ func TestCancelAllExchangeOrders(t *testing.T) {
 		"-")
 	var orderCancellation = &order.Cancel{
 		ID:            "1",
-		WalletAddress: testdata.BitcoinTestAddress,
+		WalletAddress: bitfaker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 		AssetType:     asset.Spot,
