+++ exmo/exmo_test.go
@@ -8,8 +8,8 @@ import (
 
 	exchange "github.com/openware/irix"
 	"github.com/openware/irix/config"
+	"github.com/openware/irix/faker/bitfaker"
 	"github.com/openware/irix/portfolio/withdraw"
-	testdata "github.com/openware/irix/test"
 	"github.com/openware/pkg/asset"
 	"github.com/openware/pkg/common"
 	"github.com/openware/pkg/currency"
@@ -333,7 +333,7 @@ func TestCancelExchangeOrder(t *testing.T) {
 	currencyPair := currency.NewPair(currency.LTC, currency.BTC)
 	var orderCancellation = &order.Cancel{
 		ID:            "1",
-		WalletAddress: testdata.BitcoinTestAddress,
+		WalletAddress: bitfaker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 		AssetType:     asset.Spot,
@@ -356,7 +356,7 @@ func TestCancelAllExchangeOrders(t *testing.T) {
 	currencyPair := currency.NewPair(currency.LTC, currency.BTC)
 	var orderCancellation = &order.Cancel{
 		ID:            "1",
-		WalletAddress: testdata.BitcoinTestAddress,
+		WalletAddress: bitfaker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 		AssetType:     asset.Spot,
@@ -396,7 +396,7 @@ func TestWithdraw(t *testing.T) {
 		Currency:    currency.BTC,
 		Description: "WITHDRAW IT ALL",
 		Crypto: withdraw.CryptoRequest{
-			Address: testdata.BitcoinTestAddress,
+			Address: bitfaker.BitcoinTestAddress,
 		},
 	}
