+++ faker/bitfaker/bitcoin.go
@@ -1,3 +1,3 @@
-package test
+package bitfaker
 
 const BitcoinTestAddress = "1MK6jowLN2kC5RrWoHMVqFi2Gq6yQjHQKm"
