+++ okcoin/okcoin_test.go
@@ -12,11 +12,11 @@ import (
 	"github.com/gorilla/websocket"
 	exchange "github.com/openware/irix"
 	"github.com/openware/irix/config"
+	"github.com/openware/irix/faker/bitfaker"
 	"github.com/openware/irix/okgroup"
 	"github.com/openware/irix/portfolio/withdraw"
 	"github.com/openware/irix/sharedtestvalues"
 	"github.com/openware/irix/stream"
-	testdata "github.com/openware/irix/test"
 	"github.com/openware/pkg/asset"
 	"github.com/openware/pkg/common"
 	"github.com/openware/pkg/currency"
@@ -147,7 +147,7 @@ func TestAccountWithdrawRequest(t *testing.T) {
 		Currency:    currency.BTC.String(),
 		TradePwd:    "1234",
 		Destination: 4,
-		ToAddress:   testdata.BitcoinTestAddress,
+		ToAddress:   bitfaker.BitcoinTestAddress,
 		Fee:         1,
 	}
 	_, err := o.AccountWithdraw(request)
@@ -986,7 +986,7 @@ func TestCancelExchangeOrder(t *testing.T) {
 	currencyPair := currency.NewPair(currency.LTC, currency.BTC)
 	var orderCancellation = order.Cancel{
 		ID:            "1",
-		WalletAddress: testdata.BitcoinTestAddress,
+		WalletAddress: bitfaker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 	}
@@ -1001,7 +1001,7 @@ func TestCancelAllExchangeOrders(t *testing.T) {
 	currencyPair := currency.NewPair(currency.LTC, currency.BTC)
 	var orderCancellation = order.Cancel{
 		ID:            "1",
-		WalletAddress: testdata.BitcoinTestAddress,
+		WalletAddress: bitfaker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 	}
@@ -1034,7 +1034,7 @@ func TestWithdraw(t *testing.T) {
 
 	withdrawCryptoRequest := withdraw.Request{
 		Crypto: withdraw.CryptoRequest{
-			Address:   testdata.BitcoinTestAddress,
+			Address:   bitfaker.BitcoinTestAddress,
 			FeeAmount: 1,
 		},
 		Amount:        -1,
